+++ packages/nasi-lemak-react-types/src/Rx.ts
@@ -6,12 +6,20 @@
 
 // @barrel export all
 
-import { Types } from "nasi";
+import { assert, Option, Types } from "nasi";
 import {
   Observable,
   ObservableInput,
+  OperatorFunction,
   partition as rxpartition,
+  Subject,
 } from "rxjs";
+import { declareAsStable, Dispatch as StableDispatch } from "./Stable";
+
+export interface IObservableDispatch<TAction> {
+  dispatch: StableDispatch<TAction>;
+  close: () => void;
+}
 
 /**
  * A special case of `partition` in RxJS where the predicate is actually a
@@ -22,4 +30,25 @@ export function partition<U, T extends U>(
   predicate: (value: U) => value is T,
 ): [ Observable<T>, Observable<Types.Not<U, T>> ] {
   return rxpartition<any>(source, predicate);
+}
+
+export function asDispatch<InputAction, OutputAction>(
+  output: StableDispatch<OutputAction>,
+  operator: OperatorFunction<InputAction, OutputAction>,
+): IObservableDispatch<InputAction> {
+
+  const input$ = new Subject<InputAction>();
+
+  input$.pipe(operator).subscribe({
+    next: output,
+  });
+
+  return {
+    close: () => input$.complete(),
+    dispatch: declareAsStable((action) => {
+      assert(Option.isSome, action);
+      input$.next(action);
+    }),
+  };
+
 }